Update auto_capture_yt_endscreen_videowall.js
i decided to tie the timestamp factor to the data (i wish i thought of it earlier but anyway.) this is more rational and avoids potential problem. the thing now, is to handle the existing db (if any) that is already created. we may either delete it, or convert it. so, before you deploy this new code, please refer to readme.md for how to delete/convert the existing data.